Tenant Petition info 3
git-tfs-id: [http://ev-tfs-dev:8080/tfs/CA_State_Collection/]$/OAKRAP/RAP/MAIN;C39

+++ 2.APPSERVER/FinOT.Core/DataModels/TenantPetition.cs
@@ -68,7 +68,7 @@ namespace RAP.Core.DataModels
             }
         }
         public DateTime MoveInDate { get; set; }
-        public double InitalRent { get; set; }
+        public decimal InitalRent { get; set; }
         public bool bRAPNoticeGiven { get; set; }
         public DateTime RAPNoticeGivenDate { get; set; }
         public bool bRentControlledByAgency { get; set; }
